NullValueApi: changed nullbitmasks from int8_t to int64_t and also refactor of BlockBase, ColumnBase and Table to work with new API and int64_t bitmasks

+++ dropdbase/IColumn.h
@@ -17,7 +17,7 @@ public:
     virtual void InsertNullData(int length) = 0;
     virtual float GetInitAvg() const = 0;
     virtual bool GetInitAvgIsSet() const = 0;
-    virtual std::pair<int8_t*, size_t> GetNullBitMaskForBlock(size_t blockIndex) = 0;
+    virtual std::pair<int64_t*, size_t> GetNullBitMaskForBlock(size_t blockIndex) = 0;
     virtual bool GetIsNullable() const = 0;
     virtual void SetIsNullable(bool isNullable) = 0;
     virtual bool GetIsUnique() const = 0;
